move GeomBase
